\name{biozones}
\docType{data}
\alias{biozones}
\title{Ammonite biozone ages for the Upper Cretaceous}
\description{
 This dataset provides the lower and upper ages of ammonite biozones between the Turonian and the Maastrichtian.
}
\usage{data(biozones)}
\format{A matrix of 50 biozonal ages.}
\source{

Hardenbol, J., Thierry, J., Farley, M. B., Jacquin, T., Graciansky, P.-C., and Vail, P. R., 1998, Mesozoic and Cenozoic sequence chronostratigraphic framework of European basins, in Graciansky, P.-C., Hardenbol, J., Jacquin, T., and Vail, P. R., eds., Mesozoic and Cenozoic sequence stratigraphy of European basins, Volume 60: Tulisa, OK, Society of Economic Palaeontologists and Mineralogists (SEPM) Special Publications, p. 3-14.

}
\keyword{datasets}