\name{size}
\docType{data}
\alias{size}
\title{Body-size data from trilobite specimens}
\description{

This dataset contains the measurements of trilobite specimens from the Cambrian and Ordovician. Each column represents a geological Stage.

}
\usage{data(timescales)}
\format{A matrix of geological stages for the Cambrian and Ordovician containing measurements of trilobite specimens.}
\source{

Bell, M, 2009, unpublished PhD Thesis, University of Bristol.

}

\keyword{datasets}