\name{timescales}
\docType{data}
\alias{timescales}
\title{Geological time scale from Harland et al., (2012)}
\description{

This dataset provides the geological ages for the entire geological timescale from Gradstein (2012) along with the RGB colour values used in the current version.
}
\usage{data(timescales)}
\format{A matrix of first and last datums along with RBG values for each geological time period.}
\source{

Gradstein, F. M., Ogg, J. M., and Schmitz, M., 2012, A geologic time scale, Boston, USA, Elsevier.

}

\keyword{datasets}